major structure updates to the guardrails api
